GL: Split drawText into its own file et al
